tiflibrary: XmlTvParser: Pass the correct Epg Id value
One could think that since id is a int, it could be represented the same
as a long. However, when autoboxed, the object is not the same at all
and it reflects in a Map.

In these cases, cast the id as a long in order to retrieve the list
of programs for a channel.

+++ tiflibrary/src/main/java/com/google/android/media/tv/companionlibrary/xmltv/XmlTvParser.java
@@ -549,7 +549,7 @@ public class XmlTvParser {
          * @return A list of programs that belong to that id.
          */
         public List<Program> getPrograms(int id) {
-            return mProgramMap.get(id);
+            return mProgramMap.get((long) id);
         }
     }
